Create st_mapWarp
Drive a warp with an st_map. The roto is where you want the warp to go and the transform is how you want to warp the image. You can also effect the image based on color. Red moves the images left and right and green moves the image up and down.