Create keyingTemplate1
Here is a template for keying that I use all the time. The core matte can be made anyway you want but if you are given roto or have time to do your own that works the best. Feel free to use a keyer to make a core matte but roto you know that it's a solid core. I this template I am using the IBK keyer but this can be used with any keyer works best for that area. The light can be used or not. Sometimes I find that light wrap with very low setting just takes the edge off the edges(no pun intended). Please feel free to add to this change or add a new template of your own.